+++ RELEASE-NOTES.md
@@ -5,6 +5,7 @@
 - #72 Magic parser - handling escapes and empty strings
 - #75 Completion of % and %% magics
 - #77 Relicense under Apache License
+- #78 Remove deprecated API and IJava startup vars
 
 
 ## 1.0-a5